requirements.txt
python-chess
stockfish
anthropic
python-dotenv
streamlit
plotly
requests
